extract git helper; early return
