+++ packages/tools/src/lib/build-cache/github/config.ts
@@ -1,8 +1,5 @@
-import { select } from '@clack/prompts';
 import spawn from 'nano-spawn';
-import cacheManager from '../../cacheManager.js';
 import logger from '../../logger.js';
-import { checkCancelPrompt } from '../../prompts.js';
 import { GITHUB_REPO_REGEX } from './patterns.js';
 
 export function hasGitHubToken(): boolean {
@@ -15,30 +12,10 @@ export type GitHubRepoDetails = {
   repository: string;
 };
 
-export async function detectGitHubRepoDetails(): Promise<GitHubRepoDetails | null> {
+export async function detectGitHubRepoDetails(
+  gitRemote: string
+): Promise<GitHubRepoDetails | null> {
   try {
-    let gitRemote = cacheManager.get('gitRemote');
-    if (!gitRemote) {
-      const { output: remoteOutput } = await spawn('git', ['remote']);
-      const remotes = remoteOutput.split('\n').filter(Boolean);
-      if (remotes.length > 1) {
-        gitRemote = checkCancelPrompt<string>(
-          await select({
-            message: 'Select git remote of the upstream repository:',
-            options: remotes.map((remote) => ({
-              value: remote,
-              label: remote,
-            })),
-          })
-        );
-        cacheManager.set('gitRemote', gitRemote);
-      } else if (remotes.length === 1) {
-        gitRemote = remotes[0];
-      } else {
-        logger.warn('No git remote found.');
-        return null;
-      }
-    }
     const { output: url } = await spawn('git', [
       'config',
       '--get',
